refactor(ga): re-org naming
- `googleAnalytics/` => `analytics/`
- AnalyticsLogger => AnalyticsLoggerService, as its own interface file
- `analyticsLogger.ts` => `GaLoggerService.ts`
- appropriate changes to inversify dependency ids
