+++ src/server/controllers/RedirectController.ts
@@ -4,16 +4,13 @@ import { gaTrackingId, logger } from '../config'
 import { NotFoundError } from '../util/error'
 import parseDomain from '../util/domain'
 import { DependencyIds, ERROR_404_PATH } from '../constants'
-import { AnalyticsLogger } from '../services/analyticsLogger'
+import { AnalyticsLoggerService } from '../services/interfaces/AnalyticsLoggerService'
 import { RedirectControllerInterface } from './interfaces/RedirectControllerInterface'
 import { RedirectService } from '../services/RedirectService'
 import { RedirectType } from '../services/types'
-import {
-  EventAction,
-  EventCategory,
-} from '../services/googleAnalytics/types/enum'
-import { createPageViewHit } from '../services/googleAnalytics'
-import IGaPageViewForm from '../services/googleAnalytics/types/IGaPageViewForm'
+import { EventAction, EventCategory } from '../services/analytics/types/enum'
+import { createPageViewHit } from '../services/analytics'
+import IGaPageViewForm from '../services/analytics/types/IGaPageViewForm'
 
 const TRANSITION_PATH = 'transition-page.ejs'
 
@@ -21,12 +18,12 @@ const TRANSITION_PATH = 'transition-page.ejs'
 export class RedirectController implements RedirectControllerInterface {
   private redirectService: RedirectService
 
-  private analyticsLogger: AnalyticsLogger<IGaPageViewForm>
+  private analyticsLogger: AnalyticsLoggerService<IGaPageViewForm>
 
   public constructor(
     @inject(DependencyIds.redirectService) redirectService: RedirectService,
-    @inject(DependencyIds.analyticsLogging)
-    analyticsLogger: AnalyticsLogger<IGaPageViewForm>,
+    @inject(DependencyIds.analyticsLoggerService)
+    analyticsLogger: AnalyticsLoggerService<IGaPageViewForm>,
   ) {
     this.redirectService = redirectService
     this.analyticsLogger = analyticsLogger
