+++ src/server/services/analytics/types/enum.ts
