+++ src/server/services/interfaces/AnalyticsLoggerService.ts
@@ -1,0 +1,10 @@
+export interface AnalyticsLoggerService<T> {
+  /**
+   * @param {T} pageViewHit - A page view request payload.
+   */
+  logRedirectAnalytics: (pageViewHit: T) => void
+
+  generateCookie: (
+    cookie?: string,
+  ) => [string, string, { maxAge: number }] | null
+}
